requirements.txt
requests-toolbelt==0.9.1
opencv_python==3.4.2.17
requests==2.21.0
anytree==2.6.0
fleep==1.0.1
numpy==1.16.2
pandas==0.24.2
tabulate==0.8.3
pika==1.0.0
Pillow==6.0.0
PyJWT==1.7.1
progressbar2==3.39.3
PyYAML==5.1
ffmpeg-python==0.1.17
scikit-image==0.15.0
jinja2==2.10.1
attrs==19.1.0
attr==0.3.1
behave==1.2.6
